lms-backend/user_service/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1 
PyMySQL==1.1.1
PyJWT==2.9.0
Werkzeug==3.0.3
gunicorn==22.0.0
cryptography==42.0.5
